+++ CHANGELOG.md
@@ -1,3 +1,16 @@
 ## 0.0.1
 
-Initial Release
+Initial release of the publish package, featuring Android signing key generation and signed bundle
+creation.
+
+# 1.0.0
+
+Added support for generating core and feature directories via the gen command.
+
+# 1.0.2
+
+Improved key generation process with user prompts for Common Name, Organizational Unit,
+Organization, Locality, State, Country, and Validity inputs. Defaults provided for easy use.
+
+# 1.0.4
+Fixed validity years input bug.